Added abstract class to simplify further architecture
